+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Added the ability to create objects from templates for all types of databases.
+- Added support for COLUMNSTORE INDEX in MS SQL.
 
 ### Changed